Can create new layout varible inline (Sociogram)
